Add a test script for the notebooks
This uses nbconvert to execute the notebooks and report on any
that fail. That, in turn, requires that the Datalab Jupyter
extensions be installed, but this can be acheived by running
the test script inside of the Datalab gateway Docker image.

The test script includes an example invocation for running
it inside of that image.