Support testing notebooks with Travis (#23)
This change adds a Travis build for validating notebooks.

This also adds credentials for the Travis build. That is done
by following the practice outlined
[here](https://cloud.google.com/solutions/continuous-delivery-with-travis-ci).

The basic idea is that the credentials are encrypted using a public key
for which only Travis has the corresponding private key. Before tests
are run we have Travis decrypt that key, authenticate gcloud using it,
and then delete the decrypted key.

Because we are using Travis' encrypted files feature, these tests
can only be run for pull requests that originate within the 'notebooks'
repository. For pull requests from forks, the CONTRIBUTING.md
file has been extended to document how to set up separate Travis
builds on those forks.