Change the test script to be a hidden file (#26)
The vast majority of users of this repository will simply be
using the existing notebooks as docs, rather than updating
and/or testing them.

For those users, the test script is just an implementation
detail that they do not care about. As such, we should make
that file hidden by default the same way that the `.travis.yml`
file is hidden by default.