Update the stackdriver getting started notebook. (#32)
* Rename the Monitoring folder to 'Stackdriver Monitoring'.

* Update the stackdriver getting started notebook.

The notebook is modified so that it handles projects with a large number
of GCE instances better. Instead of showing the timeseries for all
instances, the notebook now does so for instances matching a given
prefix.

Also, the output of the help is updated to reflect the underlying code
changes.